move incubating enum TimeZoneStorageStrategy to org.hibernate.type
This enum is not really an API-API, and has no usages in org.hibernate.
Application programs are expected to use o.h.annotations.TimeZoneStorageType,
and so putting this enum in o.h is pretty confusing to users.
